+++ tool/src/image_d64.rs
@@ -124,7 +124,7 @@ pub fn parse_d64_image(path: &str) -> RawImage {
 
         let densitymap = vec![DensityMapEntry {
             number_of_cells: trackbuf.len() as usize,
-            cell_size: PulseDuration(settings.cellsize as u16),
+            cell_size: PulseDuration(settings.cellsize as i32),
         }];
 
         tracks.push(RawTrack::new(
